Some warning for gcc and msvc removed
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@14471 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/int/arithmetic/argmax.hpp
@@ -156,7 +156,7 @@ namespace Gecode { namespace Int { namespace Arithmetic {
         }
       x.size(j);
 
-      if (n == y.size())
+      if (static_cast<unsigned int>(n) == y.size())
         return ES_FAILED;
       Iter::Values::Array id(d,n);
       GECODE_ME_CHECK(y.minus_v(home,id,false));
