+++ gecode/int/bin-packing/conflict-graph.hpp
@@ -210,7 +210,7 @@ namespace Gecode { namespace Int { namespace BinPacking {
         return ES_FAILED;
     }
     // Compute corresponding variables
-    ViewArray<IntView> bv(home,cur.c);
+    ViewArray<IntView> bv(home,static_cast<int>(cur.c));
     int i=0;
     for (Nodes c(cur.n); c() < nodes(); ++c)
       bv[i++] = b[c()];
@@ -337,7 +337,7 @@ namespace Gecode { namespace Int { namespace BinPacking {
     for (Nodes i(max.n); i() < nodes(); ++i)
       n[j++]=i();
     assert(j == static_cast<int>(max.c));
-    IntSet s(n,max.c);
+    IntSet s(n,static_cast<int>(max.c));
     return s;
   }
